+++ src/db/database.h
@@ -10,11 +10,11 @@ namespace db {
 		virtual ~Database() {};
 		virtual Result *list_ng() = 0;
 		virtual Result *create_ng(std::string ng_name) = 0;
-		virtual Result *delete_ng(int ng_id) = 0;
-		virtual Result *list_art(int ng_id) = 0;
-		virtual Result *create_art(int ng_id, std::string title, std::string author, std::string text) = 0;
-		virtual Result *delete_art(int ng_id, int art_id) = 0;
-		virtual Result *get_art(int ng_id, int art_id) = 0;
+		virtual Result *delete_ng(size_t ng_id) = 0;
+		virtual Result *list_art(size_t ng_id) = 0;
+		virtual Result *create_art(size_t ng_id, std::string title, std::string author, std::string text) = 0;
+		virtual Result *delete_art(size_t ng_id, size_t art_id) = 0;
+		virtual Result *get_art(size_t ng_id, size_t art_id) = 0;
 	};
 }
