+++ src/db/delete_art_query.h
@@ -7,11 +7,11 @@
 namespace db {
 	class DeleteArtQuery : public Query {
 	public:
-		DeleteArtQuery(Database& db, int ng_id, int art_id) : Query(db), newsgroup_id(ng_id), article_id(art_id) {}
+		DeleteArtQuery(Database& db, size_t ng_id, size_t art_id) : Query(db), newsgroup_id(ng_id), article_id(art_id) {}
 		Result *execute();
 	private:
-		int newsgroup_id;
-		int article_id;
+		size_t newsgroup_id;
+		size_t article_id;
 	};
 }
